Sprint 8 mods
Changed hallway door to exclude the xr. This was the only difference between it and the classroom, so hopefully it fixes it. Made teacher distance value public so it can be freely changed. Added colliders to the computer and mini game options to fix the not turning on issue. Made video not play on awake as well.